Use f32 type instead of i32.

+++ src/main.rs
@@ -1,7 +1,7 @@
 fn main() {
-    let mut sum = 0;
+    let mut sum = 0.0;
     for i in 0..10 {
-        sum += i;
+        sum += i as f32;
     }
     println!("Sum is {}", sum)
 }
